Final cleanup

+++ GraphicsApp/Application.h
@@ -14,8 +14,8 @@ class Application
 	glm::vec2 m_mousePos;
 	glm::vec2 m_lastMousePos;
 
-	float m_width = 1280;
-	float m_height = 720;
+	int m_width = 1280;
+	int m_height = 720;
 
 	std::multimap<ShaderProgram*, GameObject*> m_gameObjects;
 	Camera* m_currentCamera;
