+++ GraphicsApp/Mesh.h
@@ -11,7 +11,7 @@ protected:
 	GLuint m_vertBuffer; // Sometimes called VBO (Vertex Buffer Object)
 	GLuint m_indexBuffer; // Sometimes called IBO (Index Buffer Object)
 
-	int m_triCount;
+	size_t m_triCount;
 
 public:
 	Mesh();
